requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# LangChain ecosystem
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# Data processing
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.24.0

# Vector database
faiss-cpu>=1.7.4

# Web scraping and APIs
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
urllib3>=2.0.0

# Google APIs
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0

# Async support
aiohttp>=3.8.0

# Optional - for enhanced features
python-dateutil>=2.8.0
pytz>=2023.3